requirements.txt
# Machine learning libraries
tensorflow-gpu
matplotlib
numpy